+++ R/primitive-bracket.R
@@ -32,28 +32,40 @@
 #' @export
 #'
 #' @details
-#' # Styling options
+#' ## Styling options
 #'
 #' Below are the [theme][ggplot2::theme] options that determine the styling of
-#' this guide, which may differ depending on whether the guide is used in an
-#' axis or a legend context.
+#' this guide, which may differ depending on whether the guide is used in
+#' an axis or in a legend context.
 #'
-#' Common to both types is the following:
+#' The possible `{position}` suffixes mentioned below are `x`, `x.top`,
+#' `x.bottom`, `y`, `y.left`, `y.right`. The `theta` and `r` position suffixes
+#' in \pkg{ggplot2} are *not* obeyed in \pkg{legendry}.
 #'
-#' * `legendry.bracket` an [`<element_line>`][ggplot2::element_line] for the
-#'   line used to draw the brackets.
-#' * `legendry.backet.size` a [`<unit>`][grid::unit] setting the space afforded
-#'   to a bracket.
+#' | **Theme setting** | **Context** | **Type** | **Description** |
+#' | ----------------- | ----------- | -------- | --------------- |
+#' | `legendry.bracket` | Both | [`element_line()`] | The bracket lines themselves. |
+#' | `legendry.bracket.size` | Both | [`unit()`] | The space (in the orthogonal direction) afforded to a bracket. |
+#' | `axis.text.{position}` | Axis | [`element_text()`] | The text over brackets. |
+#' | `legend.text` | Legend | [`element_text()`] | The text over brackets. |
 #'
-#' ## As an axis guide
+#' Styling options *per level* can be set in the `levels_brackets` and
+#' `levels_text` arguments. These override theme settings.
 #'
-#' * `axis.text.{x/y}.{position}` an [`<element_text>`][ggplot2::element_text]
-#'   for the text displayed over the brackets.
+#' Styling options *per range* can be set in the [range key][key_range].
+#' The `line` and `text` prefixed properties are prioritised for the brackets
+#' and text respectively. These override theme settings and 'per level'
+#' settings.
 #'
-#' ## As a legend guide
+#' The context-agnostic alternative to using `theme()` is to use
+#' [`theme_guide()`]:
 #'
-#' * `legend.text` an [`<element_text>`][ggplot2::element_text] for the text
-#'   displayed over the brackets.
+#' ```r
+#' primitive_bracket(theme = theme_guide(
+#'   bracket = element_line(),
+#'   bracket.size = unit(5, "mm")
+#' ))
+#' ```
 #'
 #' @examples
 #' # A standard plot
@@ -143,22 +155,21 @@ PrimitiveBracket <- ggproto(
   extract_params = extract_range_params,
 
   extract_decor = function(scale, aesthetic, position, key, bracket, ...) {
-    bracket <- resolve_bracket(bracket)
-
     key <- vec_slice(key, key$.draw)
     n_keys <- nrow(key)
-
+    if (n_keys < 1) {
+      return(NULL)
+    }
+    bracket  <- resolve_bracket(bracket)
+    n_vertex <- nrow(bracket)
+    decor <- vec_rep_each(key, n_vertex)
     brackets <- vec_rep(bracket, n_keys)
-    keys <- vec_rep_each(key, nrow(bracket))
 
-    value <- brackets[, 1L] * (keys$end - keys$start) + keys$start
-
-    data_frame0(
-      !!aesthetic := value,
-      offset = brackets[, 2L],
-      group = rep(seq_len(n_keys), each = nrow(bracket)),
-      .level = keys$.level
-    )
+    decor[[aesthetic]] <-
+      brackets[, 1L] * (decor$end - decor$start) + decor$start
+    decor[["offset"]] <- brackets[, 2L]
+    decor[["group"]] <- rep(seq_len(n_keys), each = n_vertex)
+    decor[setdiff(names(decor), c("start", "end"))]
   },
 
   transform = function(self, params, coord, panel_params) {
@@ -296,9 +307,14 @@ draw_bracket <- function(decor, element, size, offset, position) {
     y <- y + unit(cos(decor$theta) * offset, "cm")
   }
 
-  id <- vec_unrep(decor$group)$times
+  id <- new_rle(decor$group)
+  props <- element_key_properties(vec_slice(decor, id$start), "line")
 
-  grob <- element_grob(element, x = x, y = y, id.lengths = id)
+  grob <- inject(element_grob(
+    element, x = x, y = y,
+    id.lengths = id$times,
+    !!!props
+  ))
   if (!is_blank(element)) {
     attr(grob, "size") <- size
   }