+++ R/primitive-ticks.R
@@ -14,44 +14,53 @@
 #' @family primitives
 #'
 #' @details
-#' # Styling options
+#' ## Styling options
 #'
 #' Below are the [theme][ggplot2::theme] options that determine the styling of
 #' this guide, which may differ depending on whether the guide is used in
 #' an axis or in a legend context.
 #'
-#' Common to both types is the following:
+#' The ticks can come in three variants: major, minor and minimal.
+#' Which variants are drawn depends on the keys: [`key_minor()`] draws major
+#' and minor ticks, whereas [`key_log()`] also has minimal ticks.
+#' Each variant has a corresponding length setting.
 #'
-#' ## As an axis guide
+#' The possible `{position}` suffixes mentioned below are `x`, `x.top`,
+#' `x.bottom`, `y`, `y.left`, `y.right`. The `theta` and `r` position suffixes
+#' in \pkg{ggplot2} are *not* obeyed in \pkg{legendry}.
 #'
-#' * `axis.ticks.{x/y}.{position}` an [`<element_line>`][ggplot2::element_line]
-#'   for major tick lines.
-#' * `axis.minor.ticks.{x/y}.{position}` an
-#'   [`<element_line>`][ggplot2::element_line] for minor tick lines.
-#' * `legendry.axis.mini.ticks` an [`<element_line>`][ggplot2::element_line]
-#'   internally inheriting from the minor ticks for the smallest ticks in e.g.
-#'   log axes.
-#' * `axis.ticks.length.{x/y}.{position}` a [`<unit>`][grid::unit] for the major
-#'   ticks length.
-#' * `axis.minor.ticks.length.{x/y}.{position}` a [`<unit>`][grid::unit] for the
-#'   minor ticks length.
-#' * `legendry.axis.mini.ticks.length` a [`<unit>`][grid::unit] internally
-#'   inheriting from the minor tick length for the smallest ticks in e.g.
-#'   log axes.
+#' | **Theme setting** | **Context** | **Type** | **Description** |
+#' | ----------------- | ----------- | -------- | --------------- |
+#' | `axis.ticks.{position}` | Axis | [`element_line()`] | Major tick lines |
+#' | `axis.ticks.length.{position}` | Axis | [`unit()`] | Major tick length |
+#' | `axis.minor.ticks.{position}` | Axis | [`element_line()`] | Minor tick lines |
+#' | `axis.minor.ticks.length.{position}` | Axis | [`unit()`] | Minor tick length |
+#' | `legendry.axis.mini.ticks` | Axis | [`element_line()`] | Minimal tick lines |
+#' | `legendry.axis.mini.ticks.length` | Axis | [`unit()`] | Minimal tick length |
+#' | `legend.ticks` | Legend | [`element_line()`] | Major tick lines |
+#' | `legend.ticks.length` | Legend | [`unit()`] | Major ticks length |
+#' | `legendry.legend.minor.ticks` | Legend | [`element_line()`] | Minor tick lines |
+#' | `legendry.legend.minor.ticks.length` | Legend | [`unit()`] | Minor ticks length |
+#' | `legendry.legend.mini.ticks` | Legend | [`element_line()`] | Minimal tick lines |
+#' | `legendry.legend.mini.ticks.length` | Legend | [`unit()`] | Minimal tick length |
 #'
-#' ## As a legend guide
+#' Styling options *per break* can be set in the [key][key_standard].
+#' The `line` and prefixed properties are prioritised for the tick lines.
+#' These override theme settings.
 #'
-#' * `legend.ticks` an [`<element_line>`][ggplot2::element_line] for major tick
-#'   lines.
-#' * `legendry.legend.minor.ticks` an [`<element_line>`][ggplot2::element_line]
-#'   for minor tick lines.
-#' * `legendry.legend.mini.ticks` an [`<element_line>`][ggplot2::element_line]
-#'   for the smallest ticks in e.g. log axes.
-#' * `legend.ticks.length` a [`<unit>`][grid::unit] for the major ticks length.
-#' * `legendry.legend.minor.ticks.length` a [`<unit>`][grid::unit] for the
-#'   minor ticks length.
-#' * `legendry.legend.mini.ticks.length` a [`<unit>`][grid::unit] for the
-#'   smallest ticks in e.g. log axes.
+#' The context-agnostic alternative to using `theme()` is to use
+#' [`theme_guide()`]:
+#'
+#' ```r
+#' primitive_ticks(theme = theme_guide(
+#'   ticks = element_line(),
+#'   ticks.length = unit(5, "mm"),
+#'   minor.ticks = element_line(),
+#'   minor.ticks.length = unit(4, "mm"),
+#'   mini.ticks = element_line(),
+#'   mini.ticks.length = unit(3, "mm")
+#' ))
+#' ```
 #'
 #' @examples
 #' # A standard plot
@@ -202,43 +211,35 @@ draw_ticks <- function(key, element, params, position, length, offset = 0.0) {
   if (n_breaks < 1L || is_blank(element) || all(length == 0L)) {
     return(zeroGrob())
   }
+  props  <- element_key_properties(key, "line")
   length <- rep_len(length, n_breaks)
-  bidi <- c(1.0, -as.numeric(params$bidi %||% FALSE))
-  if (is_theta(position)) {
-    angle  <- rep(key$theta, each = 2L)
-    x      <- rep(key$x,     each = 2L)
-    y      <- rep(key$y,     each = 2L)
+  bidi   <- c(1.0, -as.numeric(params$bidi %||% FALSE))
 
-    length <- rep_len(length, n_breaks * 2L)
-    length <- rep(bidi, times = n_breaks) * length
-    length <- unit(length + offset, "cm")
+  # Every tick has two vertices
+  length <- rep(length, each = 2L)
+  key <- vec_rep_each(key, 2L)
+  id <- rep(2L, n_breaks)
 
-    ticks <- element_grob(
-      element,
-      x = unit(x, "npc") + sin(angle) * length,
-      y = unit(y, "npc") + cos(angle) * length,
-      id.lengths = rep(2L, n_breaks)
-    )
-    return(ticks)
-  }
-  aes <- params$aesthetic
-  aes <- switch(
-    aes, x = "x", y = "y",
-    switch(params$direction, horizontal = "x", "y")
+  # Set anchor positions
+  switch(
+    position,
+    top    = {key$y <- 0.0},
+    right  = {key$x <- 0.0},
+    bottom = {key$y <- 1.0},
+    left   = {key$x <- 1.0}
+    # theta(.sec) already has appropriate x/y values
   )
 
-  mark <- unit(rep(key[[aes]], each = 2L), "npc")
+  length <- rep(bidi, times = n_breaks) * length
+  length <- unit(length + offset, "cm")
+  theta  <- get_theta(key, position)
 
-  pos <- switch(position, top = , right = 0.0, left = , bottom = 1.0)
-  dir <- (-2.0 * pos + 1.0) * bidi
-  pos <- unit(rep(pos, 2L * n_breaks), "npc")
-  tick <- unit(rep(dir, n_breaks) * rep(length, each = 2L), "cm") + pos
+  args <- list(
+    x = unit(key$x, "npc") + sin(theta) * length,
+    y = unit(key$y, "npc") + cos(theta) * length
+  )
 
-  args <- list(x = tick, y = mark, id.lengths = rep(2L, n_breaks))
-  if (position %in% c("top", "bottom")) {
-    args <- flip_names(args)
-  }
-  inject(element_grob(element, !!!args))
+  inject(element_grob(element, id.lengths = id, !!!args, !!!props))
 }
 
 zap_tick <- function(elements, name, n) {